# OpenAPI.yaml
openapi: 3.0.0
info:
  title: store
  version: "1.0"
  description: A simple store
  contact:
    name: SecuritEase Dev
    url: "https://www.securitease.com"
    email: internal@securitease.com
  termsOfService: "https://www.securitease.com"
  license:
    name: Apache 2.0
    url: "https://www.apache.org/licenses/LICENSE-2.0.htm"
servers:
  - url: "http://localhost:8080"
    description: dev
paths:
  /order:
    get:
      summary: Get Orders
      tags:
        - store
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderPage"
      operationId: getOrders
      description: Get a page of orders
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            default: 20
        - name: sort
          in: query
          required: false
          description: Sorting criteria in the form property,(asc|desc).
          schema:
            type: string
    post:
      summary: Create Order
      tags:
        - store
      responses:
        "201":
          description: Created
      operationId: createOrder
      description: Create a new order
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  format: int64
                description:
                  type: string
  /order/{id}:
    get:
      summary: Get Order
      tags:
        - store
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
      operationId: getOrderById
      description: Get an order by ID
  /customer:
    get:
      summary: Get Customers
      tags:
        - store
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerPage"
      operationId: getCustomers
      description: Get customers, optionally filtered by a case-insensitive name substring.
      parameters:
        - name: name
          in: query
          required: false
          description: Case-insensitive substring of a word in the customer's name.
          schema:
            type: string
        - name: page
          in: query
          required: false
          description: Zero-based page number used with a name query.
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          description: Maximum number of matching customers per page.
          schema:
            type: integer
            format: int32
            minimum: 1
            default: 20
        - name: sort
          in: query
          required: false
          description: Sorting criteria in the form property,(asc|desc).
          schema:
            type: string
    post:
      summary: Create Customer
      tags:
        - store
      responses:
        "201":
          description: Created
      operationId: createCustomer
      description: Create a new customer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  format: int64
                name:
                  type: string
  /product:
    get:
      summary: Get Products
      tags:
        - store
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductPage"
      operationId: getProducts
      description: Get all products
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            default: 20
        - name: sort
          in: query
          required: false
          description: Sorting criteria in the form property,(asc|desc).
          schema:
            type: string
    post:
      summary: Create Product
      tags:
        - store
      responses:
        "201":
          description: Created
      operationId: createProduct
      description: Create a new product
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  format: int64
                description:
                  type: string
  /product/{id}:
    get:
      summary: Get Product
      tags:
        - store
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found
      operationId: getProductById
      description: Get a product by ID
components:
  schemas:
    Order:
      title: Order
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
        customer:
          type: object
          properties:
            id:
              type: integer
              format: int64
            name:
              type: string
        products:
          type: array
          items:
            $ref: "#/components/schemas/OrderProduct"
    Customer:
      title: Customer
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        orders:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
                format: int64
              description:
                type: string
                x-stoplight:
                  id: mq08x82g95zl9
    CustomerPage:
      title: Customer page
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/Customer"
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        number:
          type: integer
          format: int32
    OrderPage:
      title: Order page
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        number:
          type: integer
          format: int32
    Product:
      title: Product
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
        orderIds:
          type: array
          items:
            type: integer
            format: int64
    ProductPage:
      title: Product page
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        number:
          type: integer
          format: int32
    OrderProduct:
      title: Order product
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
tags:
  - name: store
    description: store operations